feat(macos): Port vanilla Generals to macOS
Builds and runs g_generals alongside the existing z_generals, sharing Core
and Platform/MacOS. Vanilla now reaches gameplay: menu, skirmish, units,
textures and audio.

Most of the runtime work was not new porting but carrying over fixes that
only ever landed in the Zero Hour copies of per-game files:

- CreateDirectory did not translate backslashes, so a directory literally
  named "\Users\okji\" appeared in the install root and the recursive scan
  walked the whole home directory looking for archives.
- MemoryInit lacked seven pools, including MetalSurface8, which faulted
  while creating the Metal device.
- GlobalData had no adaptive startup resolution and no forced windowed
  mode, leaving the UI laid out for one resolution and drawn at another.
- DataChunk read 16-bit map strings as 32-bit WideChar, desynchronising
  the stream so build lists and waypoints resolved to garbage.
- ddsfile kept a void* in LegacyDDSURFACEDESC2, inflating the struct past
  its size check on 64-bit so every DDS was rejected and models drew black.
- render2d applied the DirectX half-texel bias, which Metal does not need.
- GameMemory did not route Apple framework allocations to the system
  allocator, so CoreAudio took memory from the engine pools and freed it
  natively.

Also carried over: null guards in Drawable::drawUIText, 16-bit replay
characters in Recorder, the score screen no longer covering campaign
missions, and the black fill behind letterboxed video.

Audio: player slots leaked because AVAudioPlayerNode drops its completion
handler when a node is disconnected mid-buffer, and the graph was rewired
on every sound. Sounds now only rewire on a format or destination change,
and a slot is reclaimed once its buffer duration has elapsed.

+++ Generals/Code/GameEngine/Source/GameClient/GUI/GUICallbacks/Menus/LanLobbyMenu.cpp
@@ -344,7 +344,7 @@ static void playerTooltip(GameWindow *window,
 		return;
 	}
 
-	UnsignedInt playerIP = (UnsignedInt)GadgetListBoxGetItemData( window, row, col );
+	UnsignedInt playerIP = (UnsignedInt)(size_t)GadgetListBoxGetItemData( window, row, col );
 	LANPlayer *player = TheLAN->LookupPlayer(playerIP);
 	if (!player)
 	{
